+++ src/render_soft/soft_r_surf.c
@@ -521,7 +521,7 @@ void R_DrawSurfaceBlock16(void) {
         pbasesource += sourcetstep;
         lightright += lightrightstep;
         lightleft += lightleftstep;
-        prowdest = (unsigned short *)((long)prowdest + surfrowbytes);
+        prowdest = (unsigned short *)((uintptr_t)prowdest + surfrowbytes);
     }
 
     prowdestbase = prowdest;
